Extract addModifier and deleteModifier to shared module
I noticed that these two beefy methods were identical in both of these
components, so I'm moving them out to shared code to reduce duplication.
